Fix getInputVoltageX10 parameter type.
Settings divisor is stored as uint16_t, not uint8_t.

+++ workspace/TS100/inc/hardware.h
@@ -44,7 +44,7 @@ enum Orientation {
 
 uint16_t getHandleTemperature();
 uint16_t getTipRawTemp(uint8_t instant);
-uint16_t getInputVoltageX10(uint8_t divisor);
+uint16_t getInputVoltageX10(uint16_t divisor);
 uint16_t getTipInstantTemperature();
 uint8_t getTipPWM();
 void setTipPWM(uint8_t pulse);
